Implements the jupyterlab_black plugin
Implements the jupyterlab_black plugin in fairly hacky way, in essence
this dumps the active code cell into a tmp file, uses a user specified
Python interpreter[1] to run black's `format_str` on it (along with the
user specified line length argument to be passed to this), dump it back
to the same file, retrieve it and set it as current code cell text
value.

Majority of this is quite hacky, most notably the usage of terminals and
listening on the messages of the terminals; this is because I have
failed to find a way to get `fs`, `fs-extra` working. Another option I
haven't explored is maybe try to reuse the kernel (NOT the one
associated to the notebook necessarily) functionality within jupyterlab
and fuel the autoformatting part of the plugin that way.. But for now,
let's just roll with this first as I get more familiar with jupyterlab.

[1] - Since we want Python 2 support, and Black only runs on Python3.6+