Eliminate possible race condition
First, brief recap of how this plugin works, it does the following:
- Dump codecell content to tmp file with cat and heredoc;
- Execute black's `format_str` using user specified Python in a heredoc
  script, dump the formatted code into the SAME tmp file, prepend the
  STDOUT_MARKER and the codecell ID(*);
- cat the now formatted tmp file and a callback that's listening on
  (*) will take the output and strip it of (*) and dump that into the
  codecell;

What's wrong now?
If the callback that's listening on (*) is connected too early, this
will catch the output and dump the heredoc python script into the
codecell, this is clearly not what we want! So let's fix this.

We do so by separating `"${STDOUT_MARKER}${codeCellIdent}"` into
`"${STDOUT_MARKER}" + "${codeCellIdent}"`, so there is no way for the
callback to catch the heredoc.